moved artificial data database attempt to it's own directory
